app.py
import streamlit as st
import speech_recognition as sr
from langchain.llms import OpenAI
import pinecone
import pandas as pd

# Initialize Pinecone
pinecone.init(api_key="your-pinecone-api-key", environment="us-west1-gcp")
index = pinecone.Index("sql-query-index")

# Load OpenAI LLM
llm = OpenAI(temperature=0.2)

# Streamlit UI
st.title("Voice-Activated SQL Assistant")

if st.button("Start Recording"):
    r = sr.Recognizer()
    with sr.Microphone() as source:
        st.info("Listening...")
        audio = r.listen(source)

    try:
        query = r.recognize_google(audio)
        st.success(f"You said: {query}")

        # Use LangChain to process the query
        prompt = f"Convert this user query to SQL: {query}"
        sql_query = llm(prompt)
        st.code(sql_query, language='sql')

        # Simulate Pinecone similarity search
        # Assume semantic vector matching happens here

    except sr.UnknownValueError:
        st.error("Could not understand audio.")
    except sr.RequestError as e:
        st.error(f"Speech Recognition error: {e}")