requirements.txt
openai
langchain
pinecone-client
streamlit
speechrecognition
pyaudio
pandas